refactor translations
