+++ src/locals/translations/en/about-us.ts
@@ -1,0 +1,10 @@
+export const aboutUs = {
+  'aboutus.para': `Cayus was created by two engineers who grew into senior management and executive roles, then went on to become professional coaches. Long-time friends, our paths kept crossing and so did our shared ambition: to make management better at scale.
+  When large language models arrived, our founder Renaud, an experienced software leader, saw an opportunity to reimagine leadership development. We spoke with HR leaders and managers across industries, uncovering the real obstacles holding leadership training back.
+  Cayus is our answer: a pragmatic platform designed to work in the flow of day-to-day business. It clears the noise, strengthens managers instead of replacing them, and delivers measurable impact for teams.`,
+  'aboutus.goal.title': 'Our goal is simple: ',
+  'aboutus.goal.statement':
+    'Give managers the time, context, and coaching they need to lead with confidence.',
+  'aboutus.title': 'Our Story',
+  'aboutus.badge': 'About Us',
+};
