+++ src/locals/translations/en/cta.ts
@@ -1,0 +1,6 @@
+export const cta = {
+  'cta.title': 'Turn every interaction into impact.',
+  'cta.description': `Plug in Cayus AI and lead better in any industry in less time. Take the first step register for your 4 week pilot today at Cayus.ai`,
+
+  'cta.btn': 'Get started for free',
+};
