+++ src/locals/translations/en/features.ts
@@ -1,0 +1,34 @@
+const features_card1 = {
+  'feature.card.1.name': 'Scales leadership growth across the organisations',
+  'feature.card.1.description':
+    'Every manager gets an AI Chief of Staff that adopts to their style. It supports better conversations and strengthens leadership muscles over time.',
+};
+
+const features_card2 = {
+  'feature.card.2.name': 'From Prep Work to Leadership Focus',
+  'feature.card.2.description':
+    'Hours of information-chasing disappear with insights delivered in seconds. Managers open Cayus and start leading, not sifting.',
+};
+
+const features_card3 = {
+  'feature.card.3.name': 'Connects to You Reality, Not Theory.',
+  'feature.card.3.description':
+    "Cayus links with calendars, email, and systems to surface context directly from managers' workflows, keeping agendas rooted in what's really happening day to day.",
+};
+
+const features_card4 = {
+  'feature.card.4.name': "Strengthens Managers, Doesn't Replace Them.",
+  'feature.card.4.description':
+    'Cayus removes grunt work but leaves interpretation to manager. It coaches with a proven framework, not a philosophy, building sharper judgement',
+};
+
+export const features = {
+  'feature.title': 'Manage Links Like a Pro',
+  'feature.description':
+    'Cayus is a powerful link management tool that helps you shorten, track, and organize all your links in one place.',
+
+  ...features_card1,
+  ...features_card2,
+  ...features_card3,
+  ...features_card4,
+};
