+++ src/locals/translations/en/get-started.ts
@@ -1,0 +1,4 @@
+export const getStarted = {
+  'getStarted.message.error': 'Failed to send email. Try again later.',
+  'getStarted.message.success': 'Email sent successfully!',
+};
