+++ src/locals/translations/en/hero.ts
@@ -1,0 +1,6 @@
+export const hero = {
+  'hero.title': 'Execution Meets,',
+  'hero.highlight': 'Intelligence',
+  'hero.para_r1': `Meet Cayus AI, your intelligent chief of staff,`,
+  'hero.para_r2': `that lives right inside your existing workflow.`,
+};
