+++ src/locals/translations/en/problem.ts
@@ -1,0 +1,22 @@
+const problemCards = {
+  // Card 1
+  'problem.card.1.title': 'Too Much Admin Work',
+  'problem.card.1.description': '56% of managers’ time is spent on admin — not leadership.',
+
+  // Card 2
+  'problem.card.2.title': 'Training Doesn’t Stick',
+  'problem.card.2.description': 'Only 20% of training translates into behaviour change.',
+
+  // Card 3
+  'problem.card.3.title': 'Overload Hurts Retention',
+  'problem.card.3.description': 'When managers are overwhelmed, retention suffers.',
+};
+
+export const problem = {
+  'problem.badge': 'The Problem',
+  'problem.title': 'The Hidden Cost of Manager Overload',
+  'problem.subtitle':
+    'Cayus gives HR visibility and control over what really drives leadership performance.',
+
+  ...problemCards,
+};
