+++ src/locals/translations/fr/aboutUs.ts
@@ -1,0 +1,8 @@
+export const aboutUs = {
+  'aboutus.para': `Cayus a été créé par deux ingénieurs devenus cadres supérieurs puis dirigeants, avant de se reconvertir en coachs professionnels. Amis de longue date, nos chemins n’ont cessé de se croiser, tout comme notre ambition commune : améliorer le management à grande échelle.
+Lorsque les grands modèles de langage sont apparus, notre fondateur Renaud, un leader expérimenté dans le domaine du logiciel, a vu une occasion de réinventer le développement du leadership. Nous avons échangé avec des responsables RH et des managers de différents secteurs, identifiant les véritables obstacles qui freinent la formation au leadership.
+Cayus est notre réponse : une plateforme pragmatique conçue pour s’intégrer naturellement dans le quotidien professionnel. Elle élimine le superflu, renforce les managers au lieu de les remplacer et produit un impact mesurable sur les équipes.
+Notre objectif est simple : offrir aux managers le temps, le contexte et l’accompagnement dont ils ont besoin pour diriger avec confiance.Is this conversation helpful so far?`,
+  'aboutus.title': 'Notre histoire',
+  'aboutus.badge': 'À propos de nous',
+};
