+++ src/locals/translations/fr/getStarted.ts
@@ -1,0 +1,4 @@
+export const getStarted = {
+  'getStarted.message.error': 'Échec de l’envoi de l’e-mail. Veuillez réessayer plus tard.',
+  'getStarted.message.success': 'E-mail envoyé avec succès !',
+};
