+++ src/locals/translations/fr/hero.ts
@@ -1,0 +1,6 @@
+export const hero = {
+  'hero.title': 'Liens intelligents avec',
+  'hero.highlight': 'Précision',
+  'hero.para_r1': `Simplifiez la gestion de vos liens avec Cayus.`,
+  'hero.para_r2': `Raccourcissez, suivez et organisez tous vos liens en un seul endroit.`,
+};
