+++ src/locals/translations/fr/index.ts
@@ -1,0 +1,13 @@
+import { aboutUs } from './aboutUs';
+import { getStarted } from './getStarted';
+import { hero } from './hero';
+import { problem } from './problem';
+
+const frTranslation = {
+  ...getStarted,
+  ...hero,
+  ...problem,
+  ...aboutUs,
+};
+
+export default frTranslation;
