+++ src/locals/translations/fr/problem.ts
@@ -1,0 +1,23 @@
+const problemCards = {
+  // Card 1
+  'problem.card.1.title': 'Trop de tâches admin',
+  'problem.card.1.description':
+    '56 % du temps des managers est consacré à l’administratif — pas au leadership.',
+
+  // Card 2
+  'problem.card.2.title': 'La formation ne tient pas',
+  'problem.card.2.description':
+    'Seuls 20 % des formations se traduisent par un changement de comportement.',
+
+  // Card 3
+  'problem.card.3.title': 'La surcharge nuit à la rétention',
+  'problem.card.3.description': 'Quand les managers sont dépassés, la rétention en souffre.',
+};
+
+export const problem = {
+  'problem.badge': 'Le problème',
+  'problem.title': 'Le coût caché de la surcharge des managers',
+  'problem.subtitle':
+    'Cayus offre aux RH de la visibilité et du contrôle sur ce qui influence réellement la performance managériale.',
+  ...problemCards,
+};
